Rename Unshelled to SansShell (#2)
* Rename Unshelled to SansShell

A former colleague at Google coined the very catchy name "Unshelled" for
a non-interactive local host agent.  I spoke with some former colleagues
about open sourcing that version, and everyone seemed amenable to the
idea.  I expected we would collaborate on it to make it suitable for use
outside Google, so I started talking about the idea internally at
Snowflake as Unshelled.  A few engineers left Google, the open sourcing
effort stalled, and it became clear it would not be released soon enough
for our needs.

I still expected that once both were released, we would work to converge
them, so I began writing an interim version at Snowflake and named it
Unshelled.  Before releasing it, I reached out to my former colleague
who coined the name, to ensure that they were okay with me using the
name, with this being the only public version if their version was never
released, and that if they did release theirs, at worst we'd converge
them.  Regrettably, they asked that I not use that name.  Out of
professional courtesy, I have agreed.

I’ve also sadly come to realize that they’re not really interested in
collaborating on a single open source project, and therefore these two
versions are unlikely to converge.  Having two open source projects with
the same name that bare little resemblance and are unlikely to converge
would be the worst outcome.

After considering many alternatives, I've settled on SansShell as the
new name.  It keeps close to the spirit of the original name, and has a
nice French twist in the theme of Snowflake project names.

* go mod tidy with go1.17.1: Now with more require