Add some spacing in ProxyConn for consistency.
Invoke: Make sure the channel gets drained early returns. Also range over it.

InvokeMany:

Document the channel has to be drained or we leak resources.
Turn the streamId map into ProxyRet's so it's simpler to manage.

Make sure we break on EOF on the stream.

Add logic in close processing such that a non-OK close on a steam we've already had data is bad.
Just abort at that point and error the remaining targets.

Add a default case if we get something other that StreamData or ServerClose.