Proxy implementation for Invoke/InvokeMany (#32)
* Update to include proxy-server

* Didn't include process. Add it in

* Make proxy an important flag

* Print the server map on startup so we can validate everything we expect is loaded.

* First pass impl for simple unary case with Invoke.

Remove ProxyMany (it doesn't make sense compared to the complexity it adds).

Make sure Dial forces at least one target.

* Comment update

* Include new protoc plugin binary

* A protoc plugin which adds OneMany methods for each service.

Only does unary at the moment.

Does via embedding of the original and extending. This way users can
create a FooProxyClient and it "just works" for original methods.

* Minor cleanups for clarify

* Convert to using proxy impl

Instead of N args passed through Execute for subcommands create a struct.
Includes the connection, if it's setup for N targets and the output file to use.

Plumb State struct through all subcommand clients.

Convert ps to use this in both forms. As InvokeMany isn't implemented it properly fails.

* Add some comments and fix a bug where the client would never see its channel closed.

* Regen after shutdown fixes

* Handle channels correctly (using ok test).

Make sure return type is allocated before Any tries to reset it in unmarshal.

Mark channels as receive only.

* Handle channels correctly with ok test and not select.

* One fix in Invoke where code/msg might do a nil deref.

ImplementInvoke OneMany.

First pass as Invoke is the degenerate case of this so we'll rewrite it in terms of a channel shortly.

* Slight revamp for usability.

The NewFooProxy version always requires a *ProxyConn since it has
methods that ClientConnInterface doesn't have.

This means proxy.Dial* should return a *ProxyConn as well.
It can still be a pass through for the non-proxy case since Invoke/NewStream can handle that.

Adjust util ExecuteState.

Also fix bug where we were reusing the typed return in the generated code which meant 2 PS at the same time showed a truncated one.

* Add a comment explaining why we're embedding the client struct as well.

This way the grpc plugin can generate the normal methods and we only have
to generate the OneMany versions.

* Implement Invoke in terms of InvokeMany since it's the degenerate case.

* Add note about installing the new plugin

* Don't need a nonces map. We just need to track server ids after we validate

* Fix error typo

* Document the calling expectations to drain the channel.

Regen.

* Simplify logic so less indentation.

Also range over the channel.

* Add some spacing in ProxyConn for consistency.

Invoke: Make sure the channel gets drained early returns. Also range over it.

InvokeMany:

Document the channel has to be drained or we leak resources.
Turn the streamId map into ProxyRet's so it's simpler to manage.

Make sure we break on EOF on the stream.

Add logic in close processing such that a non-OK close on a steam we've already had data is bad.
Just abort at that point and error the remaining targets.

Add a default case if we get something other that StreamData or ServerClose.

* Update comment. Not doing N:N

* Move output prefixes into a list instead of N check || check

* Remove a 2nd check we already did.

* Refactor XXXOneMany so if it's called with a single target it just handles it internally.

Calls Invoke instead of InvokeOneMany.
Amusingly that calls InvokeOneMany in the single case...
Then take the return and create a channel to push it down then close.

Redo proxy code where a few methods on ProxyConn are needed to support this.

* Revamp for FooOneMany supporting single targets seamlessly.

* Set a default for --targets

* Remove MultipleTargets as no longer needed.

Also migrate os.File to interface io.Writer instead

* Move Invoke case inside the go routine and deduplicate a channel create