Add code in Recv to handle the normal grpc.ClientStream calling here.
This can happen as we return this when creating a streaming RPC client but then allow it to be used in OneMany context.

So in this case detect it's a direct connection and call RecvMsg with the real return type.

This works because RecvMsg is the grpc one, not the proxy one.

Need to embed the ProxyConn inside the stream object so we can detect direct status.

Regen all the files and rip out the incorrect code in proxy RecvMsg where I originally thought this happened.
Turns out the call chain here is complex.