Add end-end integration test along with fixes (#38)
* Add code in Recv to handle the normal grpc.ClientStream calling here.

This can happen as we return this when creating a streaming RPC client but then allow it to be used in OneMany context.

So in this case detect it's a direct connection and call RecvMsg with the real return type.

This works because RecvMsg is the grpc one, not the proxy one.

Need to embed the ProxyConn inside the stream object so we can detect direct status.

Regen all the files and rip out the incorrect code in proxy RecvMsg where I originally thought this happened.
Turns out the call chain here is complex.

* Implement proxy for Ansible

* Implement proxy for exec

* Implement proxy for healthcheck

* Implement proxy for localfile

* Implement proxy for package

* Fix package list to not care about stderr (it can print this while waiting for the lock).

* Cleanup error handling here to be consistent with other commands

* An end-end integration test that fires up proxy and server.

Then executes all of the commands we can (sans Java for now).

Accounts for various problems running the same command against the same host twice (ptrace generally).

* go mod tidy

* Make a basic policy instead of relying on modifying the checked in one.

* Do Send/Recv in parallel for Sum/Stat as a bidi stream can block sends.

i.e. if we block because the other end has stopped receiving potentially by being blocked sending to us.