Add tail and immutable support to stat (#40)
* API features/changes for LocalFile:

Change Read to take a one_of indicating reading a file or tailing a file.

Add List which allows you to implement ls on a directory (with optional recursion).

Add immutable options to StatReply and FileAttributes (so one can set or remove it).

* go mod tidy

* Move path testing to util and convert things to use it.

* Implement tail support and update tests.

Also add tests for a few missing cases we can without service mocks.

* Update comment on immutability.

* Move stat support to os specific implementations.

This way immutable state can be abstracted out as OS/X and linux do it differently.

In addition Linux requires a 4.11+ kernel to do this via a syscall. Otherwise you need an ioctl...

Unfortunately no way to unit test immutable as it requires root. Will add to end-end.

* Add immutable state to stat output

* Add tests for immutable files and tail.

Also suppress terminated message for process deaths by using SIGINT.

Make sure this aborts for non-linux.

Move policy into $LOGS and don't delete it on each run.

Make job terminations quiet to avoid noise.

* Fix disown since shell isn't tracking the sudo -b, only the proxy-server

* Adjust constant to be more clear and actually use it.

* Update a comment

* Update function comment to be correct.

* Move disown and document it better

* Fix immutable testing

* Use the same stat to check that we'll be using server side (it isn't always os.Stat)

* Move immutable from direct stat as well.

* Add test case and handling for not sending a valid request.

* Implement cancel correctly for tail.

We can't poll it turns out because a file is always ready to read even at EOF. It'll just return EOF over and over..

So we have to do an old school sleep loop so we can break out to check cancel state.

Update unit tests to accomodate.