In the direct case if we get an error we need to return an error over the channel and then make sure future calls give EOF (as the proxy would do).
i.e. using the proxy methods implies acting as the proxy would. So we can't just return the real error on Recv.

Instead we have to push it into the Error fields and then construct an EOF entry as well.

Otherwise continued calls on RecvMsg will just keep returning the same error.