Add tests for proxy client code (#51)
* Refactor build step to use build.go

Move paths to use bin/

Redo testing to force coverage and validate on paths we care about we have minimum coverage.

* Remove proxy/proxy from required coverage for now (no tests).

Leave TODO. Next CL

* Add tsan checking separate from coverage

* Attempt to run integration tests

* Remove integration for now. needs work

* In the direct case if we get an error we need to return an error over the channel and then make sure future calls give EOF (as the proxy would do).

i.e. using the proxy methods implies acting as the proxy would. So we can't just return the real error on Recv.

Instead we have to push it into the Error fields and then construct an EOF entry as well.

Otherwise continued calls on RecvMsg will just keep returning the same error.

* regen for updated generator

* Rename to match other generation targets with protos

* Regenerate proxy/testsdata and add missing services

* Move the testdata server into here
along with helpers for starting them up.

* Revert "Move the testdata server into here"

This reverts commit 62f5cde8ef3167234859b877e3782eefc2e80d67.

* Revert "Revert "Move the testdata server into here""

This reverts commit 5f6226fb88d522c7b6148536a986760a6dfcb35e.

* Revert "Move the testdata server into here"

This reverts commit 62f5cde8ef3167234859b877e3782eefc2e80d67.

* Revert "Revert "Move the testdata server into here""

This reverts commit f1493027a72f0c61a6d0898a0211521581bd4b54.

* Server and proxy test change to add coverage using testutil

* Split coverage up a bit more and validate proxy gets through (with relaxed numbers).

* Remove extra copy of coverage that seemed to have snuck in

* Convert grpc.WithInsecure() to grpc.WithTransportCredentials(insecure.NewCredentials())

The former is deprecated and the latter is a drop-in replacement

* Make sure in the Invoke case we send client closes and then CloseSend the conn.

* Few test cleanups/logging

Also start implementation of a fake proxy so we can test error states

* Add a fake server so we can test various error cases.

Covers all of Invoke and most of streaming recv. The bits that aren't are duplicates so not worth setting up.

* Add service to directory we can ignore for tests.

Bump min coverage to 88%

* Comment out invoke sending closes+closeSend. This causes a data race in the server around the channel.