Implement a logging module.
This can safely set the logging verbosity level on a server using stdr package.

Implement basic unit tests and client/integration tests.

All client side support in sanssh to do this on the proxy itself.

Wire into proxy as an RPC it'll serve in addition to /Proxy.Proxy
This means we need to add interceptors for unary RPCs here now too.