Fixup case where timeout < Dial timeout causing all targets to abort.
The default Dial timeout is 20s. Generally this doesn't matter as the real connect happens when you make an RPC on the connection.
In that case your context timeout controls things anyways and will abort in Ns correctly.

But...since we cancel() on timeouts in the Dial/NewStream path this rolls all the way up and aborts the entire session.

Instead remove the cancels (not needed) and instead push NewTargetStream into a go routine.
Then any issue on the stream can simply be reported in band instead of an overall rollup failure for the whole thing.