Instead of binding flags in packages export the needed vars.
Then package main in the servers can bind them as needed to flags. Or even just hard code values.

In either case we leave it up to users how to handle this vs making it difficult with a buried flag.

Flag defaults can be changed but you have to look them up and then change the Flag struct returned.