Move all connection logic inside Run() and simply do validation only up front.
When creating a new steam all we really need to do is validate the Method and return an ID.

This allows clients to start pushing requests down before all connections are finished.

i.e. I healtcheck 100 hosts with 10 timing out. I can't even start the healtcheck until 20s have passed waiting on the dial timeout. Right now the NewStream is parallel but we still have to pay one timeout. This way no timeouts have to be paid to start work.

This does change the contract on authz handling. Push authz into the TargetStream instead of at the set level. Each target is fundamentally a discrete unit and treating this as a transaction is in hindsight (for me) a fallacy. Nothing is atomic here as any number of targets may fail for reasons. So OPA should validate each connection and be bound to each one as well. This also means you get better error messages in the end as an OPA failure is tied to the specific stream not a general error on the set.