Maintain a list of streams we've previously had but are now closed.
The client is not required to sync and likely isn't as this is a bidi stream.

Therefore it's valid for this sequence to happen:

->start
<-reply
<-close
->data

as the client may read at the reply point but not read close before sending data.

Close conditions are asynch in nature. The Recv() we're blocking on may return for a TCP reset at any time. Though in testing it was most often a dial problem showing immediately in NewStream and racing.