Move all connection logic inside Run() and simply do validation only up front. (#129)
When creating a new steam all we really need to do is validate the Method and return an ID.

This allows clients to start pushing requests down before all connections are finished.

i.e. I want to health check 100 hosts with 10 timing out. I can't even start the health check until 20s have passed waiting on the dial timeout. Right now the NewStream is parallel but we still have to pay one timeout. This way no timeouts have to be paid to start work. This also allows a future dial timeout to be plumbed into the protocol since we can do blocking dials now as each will be scoped to their own go routine instead of blocking dispatcher.

This does change the contract on authz handling. Pushing authz into the TargetStream instead of at the set level.

Each target is fundamentally a discrete unit and treating this as a transaction is in hindsight (for me) a fallacy. Nothing is atomic here as any number of targets may fail for reasons. So OPA should validate each connection and be bound to each one as well. This also means you get better error messages in the end as an OPA failure is tied to the specific stream not a general error on the set. That has a bonus of being much easier to debug. Think an OPA policy which binds hosts to groups and you include one for a group you're not in. Today you get a generic set error of failing OPA without a good idea of which host triggered it, whereas with this change you'll see the specific error on the specific stream.

* Proxy client shouldn't error if the error it got from a send was EOF.

* Make sure in the authz case we force that error to be returned and not possibly a context one.

Add tests to get coverage up.

* Simplify end of Proxy.Proxy. It's just returning an error.

* Update comment about auth checks

* Fix wrong assumption about streaming errors.

When the stream has an error, targets are closed with context errors and Recv() on the stream returns its error.

Use CloseWith in a few places to avoid overlap with logic and potential deadlock.

Change var to ReqBufferSize to reflect usage better.

* Rework stream setup in client.

Now that the server returns quicker it's possible for this to happen:

reply 0
reply 1
close 1
reply 2

i.e. we get a close before all replies. Close needs to be processed and inserted into errors and removed from tracking.

* Maintain a list of streams we've previously had but are now closed.

The client is not required to sync and likely isn't as this is a bidi stream.

Therefore it's valid for this sequence to happen:

->start
<-reply
<-close
->data

as the client may read at the reply point but not read close before sending data.

Close conditions are asynch in nature. The Recv() we're blocking on may return for a TCP reset at any time. Though in testing it was most often a dial problem showing immediately in NewStream and racing.

* Add test proving parallel works.

2 working targets
1 dial error
1 timeout

All happen in parallel and don't leak onto other targets even with short (<20s) timeouts.