With the proxy pushing all connection logic after stream reply we can add a dial timeout now (#130)
Move Dial into Run() as well and plumb a timeout into here (optional).

This allow a per target dial timeout if wanted. Blocking doesn't matter since all connection logic is handled inside a per target go routine so we can safely do the separate blocking call from Dial now.

Add the ability for DialContext in TargetDialer to take additional options at calling time. This way we can slip in a WithBlock() but only in cases it's needed.

Add tests proving timeouts work and don't error in the normal case

* Rework stream setup in client.

Now that the server returns quicker it's possible for this to happen:

reply 0
reply 1
close 1
reply 2

i.e. we get a close before all replies. Close needs to be processed and inserted into errors and removed from tracking.

* Maintain a list of streams we've previously had but are now closed.

The client is not required to sync and likely isn't as this is a bidi stream.

Therefore it's valid for this sequence to happen:

->start
<-reply
<-close
->data

as the client may read at the reply point but not read close before sending data.

Close conditions are asynch in nature. The Recv() we're blocking on may return for a TCP reset at any time. Though in testing it was most often a dial problem showing immediately in NewStream and racing.

* Add test proving parallel works.

2 working targets
1 dial error
1 timeout

All happen in parallel and don't leak onto other targets even with short (<20s) timeouts.

* Add debugging to parallel work

* Errors can be different between environments. Just count errors vs healthy.