Allow loading roots of trust from a directory.
Snowflake has an internal use case where we have multiple root .pem files in a single directory. We currently concatenate together all the files for sanshell, but avoiding the concatenation makes the system slightly simpler.

We'd previously returned an error when asking for a directory, so this change makes us slightly more permissive.