Add an option for an extremely minimal web debug server (#190)
server.WithDebugPort now allows you to start a debug port in both sansshell-server and proxy-server, currently with three things.

- Trivial index page
- /debug/pprof/ and children
- /metrics

This expands the public interface of this package, so some caution is warranted. Some notable callouts:

- There's no exposed interfaces for editing what gets displayed or adding your own pages
- There's no way to reduce the number of things being displayed if someone wants to make the debug handler less revealing

We can add more Options in the future and there's always the escape hatch of someone creating their own http server, so that seems like an acceptable tradeoff.