Add a way to do command line completion in sanssh. (#186)
See steps in the readme for how to enable this. It relies on invoking sanssh for suggestions. Downstream clients can use `client.AddCommandLineCompletion` to take advantage of this.

This completion automatically registers subcommand packages and their flags. To register subcommands of subcommands, the parent subcommand needs to implement HasSubpackage. I've updated all the services to implement that interface.

There's also a PredictArgs interface that can be used to predict command line arguments of subcommands. I've experimented with adding this to some of the fdb subcommands.

Finally, I've tweaked the default fail-early behavior so that there's help text from running a command like `sanssh` or `sanssh exec` insted of a cryptic "Must set a target or a proxy" error.