Add a way to expose denial hints in the sansshell policies. (#212)
* Add a way to expose denial hints in the sansshell policies.

A common issue when interacting with complex policies is figuring out why a user doesn't have permission to run a command. It's possible to add print statements to diagnose or to grab a copy of the input and run it against a modified local version of a policy, but this tends to be time-consuming and requires knowledge of how policy evaluation works.

This change adds a new way to help with common issues by giving a way to put failure messages into the policy. Normally policies have a single `allow` boolean that gets consulted to decide whether a command is approved. I've added a way to create an additional array of strings that gets queried on a subsequent evaluation after failure.

If you add a line like the following to the rego policy

```
denial_hints [msg] {
  not allow
  msg :="you need to be allowed"
}
```

Then rejected requests will see a failure message like

```
OPA policy does not permit this request: [you need to be allowed]
```

Enabling this requires passing in an option like `opa.WithDenialHintsQuery("data.sansshell.authz.denial_hints")` when creating the OPA policy.

* Tweak proxy/server setup to use this

* Fail on missing policy

* Remove accidental comment

* Make reviewer-requested changes

* Add comment

* Correct WithMetricsPort now that Option interface has changed