Fix `fdbcli fdb kill` (#238)
Suppose we want to kill an FDB process. One can try to do it naively via sansshell in a following way but it would fail:

```
$ sanssh --targets=${IP?} fdb fdbcli kill 1.2.3.4:4500
fdbcli error: rpc error: code = Internal desc = error from running - <nil>
stdout:

stderr:
ERROR: process `1.2.3.4:4500' not recognized.
```

This is because fdbcli requires kill to be done in two phases: first we get a list of processes by running kill without arguments which fetches list of processes and caches it in the client, so kill <process> works after that. But it’s not convenient to do this via sanshell as it’s not designed for interactive commands. We can do sanssh --targets=${IP?} fdb fdbcli --exec 'kill; kill [proc]' However any kind of --exec reduces our ability to have fine-graned ACLs and auditing.

Sansshell client even documents that `fdbcli kill [proc]` will fetch process list first, however it is not implemented on the server side. This PR fixes it:

* Make help of the fdbcli kill command shorter

* Update the autogenerated protobuf files for FDB

* Prepend kill all and kill <proc> with 'kill ; ' as CLI documentation suggests

* Add -sleep flag to the fdbcli kill command

* Add some of the cSpell false-positives to a local dictionary

* Address Steven's comments - style improvements