Exit early on empty input in prefix writer.
This fixes a crash with a signature like

```
panic: runtime error: slice bounds out of range [:-1] [recovered]
        panic: runtime error: slice bounds out of range [:-1]
```

The issue is

```
	if n := bytes.LastIndex(b, []byte{'\n'}); n == len(b)-1 {
		p.start = true
		b = b[:len(b)-1]
	}
```

bytes.LastIndex returns -1 when the bytes aren't found and len(b)-1 is -1 when the array is empty, so -1==-1 and things go boom when we try to index into the empty slice.

I've added a new test to more thoroughly test the prefix writer.