Add /Exec.Exec/StreamingRun
`/Exec.Exec/StreamingRun` is a variant of `/Exec.Exec/Run` that streams back output as the command runs instead of collecting output and returning it at the end. Many commands like ping and tcpdump are commonly used in an interactive way where output is continually printed out and are badly supported by the existing Run call. Streaming the response allows continually printing out output during a call without widening attack surfaces.

StreamingRun uses the exact same request and response as Run. The only difference is that the response is a stream that may be returned many times, each time with a portion of stdout and/or stderr. The final message in the response contains a nonzero exit code if needed.

StreamingRun is a much better experience for anyone using `sanssh exec run`, but we don't want to break backwards compatibility or cause issues when preexisting policies don't allow the new calls or preexisting sansshell nodes haven't been built with the new feature. It can be enabled by passing in `--streaming` to the command or modifying the DefaultStreaming package-level variable in services/exec/client to change the defaults.