Use separate timeouts for separate batches (#289)
When operating on large numbers of nodes, it's common for a subset of nodes to be offline. The existing logic used a single timeout for the entire command. An offline node may use up the entire timeout, so we see every subsequent batch immediately fail.

Moving the timeouts to be per-batch should solve this.