Fix licensing check
Our licensing check in our integration test was running in a subshell, so the exit error code was swallowed by the subshell instead of failing the test. We now check the exit code to see if the license check succeeded.

I've also made the licensing check significantly looser. It now checks for a single line from the Apache license, which lets us ignore minor formatting differences or different copyright years.