Properly close connections in the proxy.
The proxy was doing a bunch of work to properly call ClientClose on streams when clients had no more messages, but it neglected to close the underlying ClientConn used by the stream. This leaked the stream on each outwards dial to a sansshell server, eventually accumulating many client connections all doing nothing but keepalive.

To fix this, I'm closing the connection to a server after we've finished all send/recv calls with the server and we're about to return the final ServerClose message back to the client.

This is reproducable by launching the proxy+server and making a bunch of calls

```
% go run ./cmd/sansshell-server
% go run ./cmd/proxy-server
% for f in $(seq 20); do ./sanssh --proxy=localhost:50043 --targets=localhost:50042 file read /etc/hosts& done
```

Then watch the keepalive goroutine count at http://localhost:50044/debug/pprof/goroutine?debug=1 go up and up.

I've updated the README a bit so that it has better instructions for running the proxy.