Don't fail on missing principal info when proxying identities (#385)
* Don't fail on missing principal info when proxying identities

It's valid to have a sansshell setup where communication from a user to the proxy uses an authz hook that populates principal and communication from the proxy to the server doesn't do so. MPA failed for these with "missing peer information" because we checked for principal regardless of whether we were proxying. This change fixes that logic to only fail when both proxied identity and principal are absent..

The testing strategy for adding a proxied identity to a context is a bit odd, but I don't want to widen the public interfaces for proxiedidentity yet.

* Add error check