Fix a race condition in fdbmovedata. (#395)
The sansshell builds have been slightly flaky since my https://github.com/Snowflake-Labs/sansshell/pull/388. I introduced an issue where command output might be skipped if output is generated and the command completes after `Reader()` is called and before `Next()` is called. This caused TestFDBMoveData, the only test without a sleep in the command, to fail ~5% of the time.

The fix is to check if there's any pending data on the `next` channel before returning the final error.