Reduce telemetry logging. (#458)
The current telemetry handling tends to be overly verbose. We log both here and during authorization (https://github.com/Snowflake-Labs/sansshell/pull/331), so any RPC request tends to have both a "rpc started" and a "rpc authorized" message. We also log peer information and method information within the authorization, so it's redundant to have them both at the top level and in the authorization.

Omitting the peer and method attributes will make them dissapear from other logs. It's still possible to correlate across logs with `sansshell-trace-id`.

I've left in the log messages at V(1) so that they're there if anybody needs to debug.